+++ Sources/ProjectDrivers/BazelProjectDriver.swift
@@ -103,7 +103,7 @@ public class BazelProjectDriver: ProjectDriver {
         contextLogger.debug("Build file written to \(buildPath)")
 
         if configuration.outputFormat.supportsAuxiliaryOutput {
-            let asterisk = colorize("*", .boldGreen)
+            let asterisk = Logger.colorize("*", .boldGreen)
             logger.info("\(asterisk) Building...")
         }